refactor(dpp): split shielded builder into per-function files with tests
Split the monolithic builder/mod.rs (639 lines) into individual files
per builder function, keeping shared types and internal helpers in mod.rs.
Added 9 tests covering error paths and happy paths for all 5 builders.
